# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Projetos de Energia Solar",
        "description": "Documentação da API para gestão de projetos de energia solar",
        "contact": {
            "email": "contato@seudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9000",
            "description": "Servidor API principal"
        }
    ],
    "paths": {
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Obter detalhes de um projeto específico",
                "description": "Retorna os detalhes de um projeto pelo seu ID",
                "operationId": "b3c88a543d66b9360023ce1fc0b614f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do projeto que deseja obter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do projeto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do projeto",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cliente_id": {
                                            "description": "ID do cliente associado",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "description": "Nome do projeto",
                                            "type": "string",
                                            "example": "Projeto Solar João"
                                        },
                                        "uf": {
                                            "description": "Unidade Federativa do projeto",
                                            "type": "string",
                                            "example": "SP"
                                        },
                                        "tipo_instalacao": {
                                            "description": "Tipo de instalação",
                                            "type": "string",
                                            "example": "Laje"
                                        },
                                        "equipamentos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "equipamento_id": {
                                                        "description": "ID do equipamento",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "quantidade": {
                                                        "description": "Quantidade do equipamento",
                                                        "type": "integer",
                                                        "example": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projeto não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Remover um projeto",
                "description": "Remove um projeto específico pelo ID",
                "operationId": "8300d03aa800d21b69cc5cf7610793e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do projeto que deseja remover",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Projeto removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projeto removido com sucesso."
                                        },
                                        "project": {
                                            "description": "Informações do projeto removido",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}